Mast a few changes and will continue to Pick up after learning DOM manipulation.
The game works as intended but I cannot get the game() function to work.
The purpose of that function is to add a counter and declare a winner after 5 rounds.
I checked out the exaples of what other people completed but it seems like thay all
had some form of DOM manipulation(have NOT mastered it yet)
I will continue to work on this in the future to make sure the program is updated